revised scripts
